Reorganized/bug fix for CampusInfo
Reorganized the custom campus components for CampusInfo and fixed the CF Everywhere bugs with button display and This Week section title
